+++ src/main/java/com/example/demo/common/auth/infrastructure/resolver/CurrentUser.java
@@ -1,4 +1,4 @@
-package com.example.demo.common.auth.resolver;
+package com.example.demo.common.auth.infrastructure.resolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
