+++ src/main/java/com/example/demo/common/auth/service/AuthService.java
@@ -1,8 +1,8 @@
 package com.example.demo.common.auth.service;
 
-import com.example.demo.common.auth.dto.TokenRequest;
-import com.example.demo.common.auth.dto.TokenResponse;
-import com.example.demo.common.auth.jwt.JwtTokenProvider;
+import com.example.demo.common.auth.controller.dto.TokenRequest;
+import com.example.demo.common.auth.controller.dto.TokenResponse;
+import com.example.demo.common.auth.infrastructure.jwt.JwtTokenProvider;
 import com.example.demo.user.domain.User;
 import com.example.demo.user.service.UserRepository;
 import lombok.RequiredArgsConstructor;
